chore: update to rust 2024 edition (#889)
* chore: bump rust edition to 2024

* chore: another chunk of clippy fixes

* chore: fix warning in bindings

* chore: restructure awkward unsafe blocks

* chore: reformat

* chore: clippy fix

* chore: bump rust release in docker build

+++ lact-schema/src/profiles.rs
@@ -30,14 +30,14 @@ impl ProfileWatcherState {
     }
 
     pub fn remove_process(&mut self, pid: i32) -> Option<ProfileProcessInfo> {
-        if let Some(mut info) = self.process_list.shift_remove(&pid) {
-            if let Entry::Occupied(mut entry) = self.process_names_map.entry(info.name) {
-                entry.get_mut().remove(&pid);
-                if entry.get().is_empty() {
-                    let (key, _) = entry.remove_entry();
-                    info.name = key;
-                    return Some(info);
-                }
+        if let Some(mut info) = self.process_list.shift_remove(&pid)
+            && let Entry::Occupied(mut entry) = self.process_names_map.entry(info.name)
+        {
+            entry.get_mut().remove(&pid);
+            if entry.get().is_empty() {
+                let (key, _) = entry.remove_entry();
+                info.name = key;
+                return Some(info);
             }
         }
         None
